pyproject.toml
[tool.poetry]
name = "new-words"
version = "0.1.0"
description = ""
authors = ["Alexander Bukharin <alexawm182@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
django_fsm = "^2.6"
psycopg2-binary = "^2.8"
nltk = "^3.4"
djangorestframework = "^3.10"
markdown = "^3.1"
django-filter = "^2.2"
django-cors-headers = "^3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
ipython = "^7.9"
isort = "^4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
